Update equipment.md
I changed the mk3 to the mk4 and the mk2 to mk3.  My only question is about the MK4 NFC. I'm not sure what the implications to the air gap are, but I also am not sure if they sell the mk3 anymore.  I also think the nfc can be disabled, but I haven't verified this yet.